+++ src/test/java/com/example/petbuddybackend/middleware/filter/ForceCsrfTokenGenerationFilterTest.java
@@ -1,4 +1,4 @@
-package com.example.petbuddybackend.filter;
+package com.example.petbuddybackend.middleware.filter;
 
 import jakarta.servlet.ServletException;
 import org.junit.jupiter.api.Test;
