+++ src/test/java/com/example/petbuddybackend/middleware/filter/UserRegistrationFilterTest.java
@@ -1,8 +1,7 @@
-package com.example.petbuddybackend.filter;
+package com.example.petbuddybackend.middleware.filter;
 
-import com.example.petbuddybackend.entity.user.Client;
+import com.example.petbuddybackend.middleware.filter.UserRegistrationFilter;
 import com.example.petbuddybackend.service.user.ClientService;
-import com.example.petbuddybackend.service.user.UserService;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
